Delete upgrade utils
