+++ projects/plugins/boost/app/assets/src/js/pages/getting-started/GettingStarted.svelte
@@ -8,7 +8,7 @@
 	import config, { markGetStartedComplete } from '../../stores/config';
 	import { connection } from '../../stores/connection';
 	import { recordBoostEvent } from '../../utils/analytics';
-	import { getUpgradeURL } from '../../utils/upgrade';
+	import { getUpgradeURL } from '../../utils/connection';
 
 	$: pricing = $config.pricing;
